\name{GrandParisPackage}
\alias{GrandParisPackage-package}
\docType{package}
\title{
\packageTitle{GrandParisPackage}
}
\description{
\packageDescription{GrandParisPackage}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{GrandParisPackage}
\packageIndices{GrandParisPackage}
A package to perform particle smoothing in partially observed diffusion processes}
\author{
\packageAuthor{GrandParisPackage}

Maintainer: \packageMaintainer{GrandParisPackage}
}
\references{See the article from Gloaguen et al 2018}
\keyword{SDE Smoothing}
\seealso{
SDE
}
\examples{
~~ simple examples of the most important functions ~~
}
