% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propositionModel_functions.R
\name{Prop_evalTrans}
\alias{Prop_evalTrans}
\title{Simulate a SINE process}
\usage{
Prop_evalTrans(oldParticles, newParticles, delta = 1, RW = 1,
  theta = 3.14, sigma2 = 1)
}
\arguments{
\item{oldParticles}{Starting particles}

\item{newParticles}{vector of simulation times}

\item{theta}{Theta parameter of the model, default to pi}

\item{sigma2}{observation process variance}
}
\value{
a vector same length as times
}
\description{
function that simulates a SINE model trajectory using exact algorithm
}
