% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ProposalSINEModel}
\alias{ProposalSINEModel}
\title{Proposition model}
\description{
POD with the unidimensional SDE sinus model dX_t = sin(X_t - theta)dt + dW_t
and the observation model Y_t = X_t + N(0, sigma2), the proposal model is a random walk
}
