% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINEModel_functions.R
\name{SINE_density}
\alias{SINE_density}
\title{compute density  in the SINE model}
\usage{
SINE_density(theta = pi, X, times, MC.size = 30, log = F)
}
\arguments{
\item{theta}{Theta parameter of the model, default to pi}

\item{X}{Process values}

\item{times}{vector of simulation times}

\item{MC.size}{Monte Carlo effort to compute density}

\item{log}{should the log be returned}
}
\value{
a vector having length = length(X) - 1 giving the density of each segment
}
\description{
compute density  in the SINE model
}
