% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINEModel_functions.R
\name{SINE_logDensityGradient}
\alias{SINE_logDensityGradient}
\title{compute logdensity gradient in the SINE model}
\usage{
SINE_logDensityGradient(X, times, theta = pi, MC.size = 30)
}
\arguments{
\item{X}{Process values}

\item{times}{vector of simulation times}

\item{theta}{Theta parameter of the model, default to pi}

\item{MC.size}{Monte Carlo effort to compute density}
}
\value{
a vector having length = length(X) - 1 giving the density of each segment
}
\description{
compute logdensity gradient in the SINE model
}
