% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SINE_POD}
\alias{SINE_POD}
\title{Class of partially diffusion process ruled by a SINE model}
\description{
POD with the unidimensional SDE sinus model dX_t = sin(X_t - theta)dt + dW_t
and the observation model Y_t = X_t + N(0, sigma2)
}
