\name{GrandParisPackage-package}
\alias{GrandParisPackage-package}
\alias{GrandParisPackage}
\docType{package}
\title{
\packageTitle{GrandParisPackage}
}
\description{
\packageDescription{GrandParisPackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GrandParisPackage}
\packageIndices{GrandParisPackage}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{GrandParisPackage}

Maintainer: \packageMaintainer{GrandParisPackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
